package renaming analytics->spin
